+++ src/System.Linq.Expressions/src/System/Dynamic/Utils/ExpressionVisitorUtils.cs
@@ -6,7 +6,7 @@ using System.Linq.Expressions;
 
 namespace System.Dynamic.Utils
 {
-    internal static partial class ExpressionVisitorUtils
+    internal static class ExpressionVisitorUtils
     {
         public static Expression[] VisitBlockExpressions(ExpressionVisitor visitor, BlockExpression block)
         {
@@ -51,6 +51,31 @@ namespace System.Dynamic.Utils
                     for (int j = 0; j < i; j++)
                     {
                         newNodes[j] = nodes.GetParameter(j);
+                    }
+                    newNodes[i] = node;
+                }
+            }
+            return newNodes;
+        }
+
+        public static Expression[] VisitArguments(ExpressionVisitor visitor, IArgumentProvider nodes)
+        {
+            Expression[] newNodes = null;
+            for (int i = 0, n = nodes.ArgumentCount; i < n; i++)
+            {
+                Expression curNode = nodes.GetArgument(i);
+                Expression node = visitor.Visit(curNode);
+
+                if (newNodes != null)
+                {
+                    newNodes[i] = node;
+                }
+                else if (!object.ReferenceEquals(node, curNode))
+                {
+                    newNodes = new Expression[n];
+                    for (int j = 0; j < i; j++)
+                    {
+                        newNodes[j] = nodes.GetArgument(j);
                     }
                     newNodes[i] = node;
                 }
